AtExit and ProcessorCounter are now in Utility module.
