+++ lib/concurrent.rb
@@ -1,7 +1,6 @@
 require 'concurrent/version'
 
 require 'concurrent/synchronization'
-require 'concurrent/at_exit'
 
 require 'concurrent/configuration'
