+++ lib/concurrent/utilities.rb
@@ -1,4 +1,5 @@
+# DEPRECATED Remove this file at v1.0
 require 'concurrent/utility/monotonic_time'
-require 'concurrent/utility/processor_count'
+require 'concurrent/utility/processor_counter'
 require 'concurrent/utility/timeout'
 require 'concurrent/utility/timer'
