+++ lib/concurrent/utility/at_exit.rb
